[DDW-722] Code cleanup
