+++ source/common/types/hardware-wallets.types.ts
@@ -1,3 +1,5 @@
+import { BridgeInfo, Device as TrezorDevice, UdevInfo } from 'trezor-connect';
+
 export type BIP32Path = Array<number>;
 export type LedgerModel = 'nanoS' | 'nanoX';
 export type TrezorModel = '1' | 'T';
@@ -31,13 +33,10 @@ export const DeviceModels: {
   TREZOR_ONE: '1',
   TREZOR_T: 'T',
 };
-export const DeviceTypes: {
-  LEDGER: DeviceType;
-  TREZOR: DeviceType;
-} = {
-  LEDGER: 'ledger',
-  TREZOR: 'trezor',
-};
+export enum DeviceTypes {
+  LEDGER = 'ledger',
+  TREZOR = 'trezor',
+}
 export const DeviceEvents: {
   CONNECT: DeviceEvent;
   CONNECT_UNACQUIRED: DeviceEvent;
@@ -296,6 +295,42 @@ export type TrezorSignTransactionResponse = {
   success: boolean;
   payload: TrezorSerializedTxPayload | TrezorRawTxPayload;
 };
+
+export type LedgerDevicePayload = {
+  disconnected: boolean;
+  deviceType: 'ledger';
+  deviceId: string | null;
+  deviceModel: string;
+  deviceName: string;
+  path: string;
+  product: string;
+};
+
+export type TrezorDevicePayload = {
+  disconnected: boolean;
+  deviceId: string;
+  deviceType: 'trezor';
+  deviceModel: TrezorDevice;
+  // e.g. "1" or "T"
+  deviceName: string;
+  path: string;
+  eventType: string;
+};
+
+export type TrezorDeviceErrorPayload = {
+  deviceType: 'trezor';
+  error?: {
+    payload: {
+      error: string;
+      bridge?: BridgeInfo;
+      udev?: UdevInfo;
+      code?: string;
+    };
+  };
+};
+
+// FIXME: This should be LedgerDevicePayload | TrezorDevicePayload | TrezorDeviceErrorPayload
+// Changing so will result in a huge refactoring on HardwareWalletsStore._changeHardwareWalletConnectionStatus
 export type HardwareWalletConnectionRequest = {
   disconnected: boolean;
   deviceType: DeviceType;
